requirements.txt
flask
flask-migrate
flask-script
flask-httpauth
flask-sqlalchemy
validate_email
ipaddress

